+++ lib/Network/MPD/Parse.hs
@@ -1,31 +1,111 @@
 module Network.MPD.Parse
-  ( getStatusField
+  ( TagField (..)
+  , ExtractedTags (..)
+  , getAllTags
+  , getStatusField
   , getStatusFieldElement
+  , SongCurrentOrNext(..)
   , getTag
-  , processSong
+  , songToTagField
   , maybePathCurrentSong
   , maybePathNextPlaylistSong
-  , headMay
-  , valueToStringMay
-  , (.=?)
-  , objectJson
+  , singleValueToString
+  , multiValueToString
   , getStatusIdInt
-  ) where
+  )
+where
 
 import qualified Network.MPD as MPD
 import Network.MPD
        ( Metadata(..), Song, PlaybackState(Stopped, Playing, Paused) )
-import Data.Aeson ( object, Key, KeyValue(..), ToJSON, Value )
-import Data.Aeson.Types ( Pair )
-import Data.Maybe ( catMaybes, fromMaybe )
+import Data.Maybe ( fromMaybe )
+import Data.List ( (!?) )
 
-{- | Extract a field from the returned MPD.Status data record.
+{- | Wrapper for the output of 'getTag', which internally uses
+'Network.MPD.sgGetTag' to retrieve @Maybe@ ['Network.MPD.Value'] that
+are then converted to @TagField@. This allows handling multi-value
+tags like multiple artists.
+-}
+data TagField = SingleTagField !(Maybe String)
+              | MultiTagField !(Maybe [String])
+  deriving (Show, Eq)
 
-Helper to extract a specific field from the
-[Network.MPD.Status](Network.MPD#Status) data record by providing the
-corresponding field label. If the input status "@st@" is /not/ @Right a@,
-indicating an error, or the field label function is not applicable, it
-returns @Nothing@.
+{- | Store the parsed output of 'getTag'.
+
+Each field represents a supported MPD tag.
+-}
+data ExtractedTags = ExtractedTags
+  { artist                     :: !TagField
+  , artistSort                 :: !TagField
+  , album                      :: !TagField
+  , albumSort                  :: !TagField
+  , albumArtist                :: !TagField
+  , albumArtistSort            :: !TagField
+  , title                      :: !TagField
+  , track                      :: !TagField
+  , name                       :: !TagField
+  , genre                      :: !TagField
+  , date                       :: !TagField
+  , originalDate               :: !TagField
+  , composer                   :: !TagField
+  , performer                  :: !TagField
+  , conductor                  :: !TagField
+  , work                       :: !TagField
+  , grouping                   :: !TagField
+  , comment                    :: !TagField
+  , disc                       :: !TagField
+  , label                      :: !TagField
+  , musicbrainz_ArtistId       :: !TagField
+  , musicbrainz_AlbumId        :: !TagField
+  , musicbrainz_AlbumartistId  :: !TagField
+  , musicbrainz_TrackId        :: !TagField
+  , musicbrainz_ReleasetrackId :: !TagField
+  , musicbrainz_WorkId         :: !TagField
+  }
+
+{- | Assign 'getTag' returned values to 'ExtractedTags'.
+
+Takes either a song @Current s@ or @Next s@, because their object
+format differs, see 'SongCurrentOrNext'.
+-}
+getAllTags :: SongCurrentOrNext -> ExtractedTags
+getAllTags s = ExtractedTags
+  { artist                     = f Artist                     s
+  , artistSort                 = f ArtistSort                 s
+  , album                      = f Album                      s
+  , albumSort                  = f AlbumSort                  s
+  , albumArtist                = f AlbumArtist                s
+  , albumArtistSort            = f AlbumArtistSort            s
+  , title                      = f Title                      s
+  , track                      = f Track                      s
+  , name                       = f Name                       s
+  , genre                      = f Genre                      s
+  , date                       = f Date                       s
+  , originalDate               = f OriginalDate               s
+  , composer                   = f Composer                   s
+  , performer                  = f Performer                  s
+  , conductor                  = f Conductor                  s
+  , work                       = f Work                       s
+  , grouping                   = f Grouping                   s
+  , comment                    = f Comment                    s
+  , disc                       = f Disc                       s
+  , label                      = f Label                      s
+  , musicbrainz_ArtistId       = f MUSICBRAINZ_ARTISTID       s
+  , musicbrainz_AlbumId        = f MUSICBRAINZ_ALBUMID        s
+  , musicbrainz_AlbumartistId  = f MUSICBRAINZ_ALBUMARTISTID  s
+  , musicbrainz_TrackId        = f MUSICBRAINZ_TRACKID        s
+  , musicbrainz_ReleasetrackId = f MUSICBRAINZ_RELEASETRACKID s
+  , musicbrainz_WorkId         = f MUSICBRAINZ_WORKID         s
+  }
+  where
+    f = getTag
+
+{- | Extract a field from the returned 'Network.MPD.Status' data record.
+
+Helper to extract a specific field from the 'Network.MPD.Status' data
+record by providing the corresponding field label. If the input status
+"@st@" is /not/ @Right a@, indicating an error, or the field label
+function is not applicable, it returns @Nothing@.
 
 ==== __Example__:
 
@@ -55,24 +135,63 @@ Just 100
 getStatusFieldElement :: MPD.Response MPD.Status -> (MPD.Status -> Maybe a) -> Maybe a
 getStatusFieldElement status item = fromMaybe Nothing $ getStatusField status item
 
-{- | @Either@ check for the returned value of 'Network.MPD.currentSong',
-then call 'processSong' or return @Nothing@.
--}
-getTag :: Metadata -> Either a (Maybe Song) -> Maybe String
-getTag t c =
-  case c of
-    Left _ -> Nothing
-    Right song -> processSong t song
+-- | Alias for the output of 'Network.MPD.currentSong'.
+type CurrentSong = MPD.Response (Maybe Song)
 
-{- | Use 'Network.MPD.sgGetTag' to extract a @tag@ from a @song@, safely
-get only the head item of the returned @Maybe@ list, then safely
-convert it to a string.
+-- | Alias for the output of 'Network.MPD.playlistInfo'.
+type NextSong = MPD.Response [Song]
+
+-- | Wrapper for 'getTag' to expect either @Maybe Song@ or
+-- @[Song]@. This simplifies 'getAllTags'.
+data SongCurrentOrNext = Current !CurrentSong
+                       | Next !NextSong
+
+-- | Retrieve @tag@, which should be one of 'Network.MPD.Metadata', from
+-- 'CurrentSong' or 'NextSong'.
+getTag :: Metadata -> SongCurrentOrNext -> TagField
+getTag tag (Current song) =
+  case song of
+    Left _ -> SingleTagField Nothing
+    Right (Just s) -> songToTagField tag s
+getTag tag (Next song) =
+  case song of
+    Right [s] -> songToTagField tag s
+    Left _    -> SingleTagField Nothing
+    _any      -> SingleTagField Nothing
+
+{- | Extract a @tag@ 'Network.MPD.Value' from 'Network.MPD.Song' using
+'Network.MPD.sgGetTag', convert the output to either @Maybe String@ or
+@Maybe [String]@ and wrap it in 'TagField'.
+
+Because 'Network.MPD.sgGetTag' returns @Maybe@ ['Network.MPD.Value']
+where @Value@ is an instance of @ByteString@ it also offers helper
+conversion functions, so convert it to @String@ if the field only
+contains a list of one value or convert all ['Network.MPD.Value'] list
+items to @String@ and return the list.
 -}
-processSong :: Metadata -> Maybe Song -> Maybe String
-processSong _ Nothing = Nothing
-processSong tag (Just song) = do
-  let tagVal = MPD.sgGetTag tag song
-  valueToStringMay =<< (headMay =<< tagVal)
+songToTagField :: Metadata -> Song -> TagField
+songToTagField tag song = tagSingleOrList (MPD.sgGetTag tag song)
+  where
+    tagSingleOrList :: Maybe [MPD.Value] -> TagField
+    tagSingleOrList val | fmap length val == Just 1 = SingleTagField $ singleValueToString $ (fromMaybe [] val) !? 0
+                        | fmap length val > Just 1 = MultiTagField $ multiValueToString val
+                        | otherwise = SingleTagField Nothing
+
+{- | Convert 'Network.MPD.Value' to @String@ within a @Maybe@ context.
+
+'MPD.sgGetTag' returns a @Maybe [Value]@. [libmpd](Network.MPD) also
+provides 'Network.MPD.toString' that can also, along with @ByteString@
+and @Text@, convert a 'Network.MPD.Value' to a @String@.
+-}
+singleValueToString :: Maybe MPD.Value -> Maybe String
+singleValueToString (Just x) = Just (MPD.toString x)
+singleValueToString Nothing = Nothing
+
+-- | Same as 'singleValueToString' but converts all @Value@s in the
+-- multi-value-tag list to @String@ and returns the list.
+multiValueToString :: Maybe [MPD.Value] -> Maybe [String]
+multiValueToString (Just x) = Just $ map MPD.toString x
+multiValueToString Nothing = Nothing
 
 {- | Get the current 'Network.MPD.Song' relative path with 'Network.MPD.sgFilePath'
 -}
@@ -93,65 +212,8 @@ maybePathNextPlaylistSong (Right [])      = Nothing
 maybePathNextPlaylistSong (Right (_:_:_)) = Nothing
 maybePathNextPlaylistSong (Right [s]) =  Just $ MPD.toString $ MPD.sgFilePath s
 
-{- | Safely get the head of a list. Same as [Safe.headMay](Safe#headMay).
--}
-headMay :: [a] -> Maybe a
-headMay []    = Nothing
-headMay (x:_) = Just x
-
-{- | Convert 'Network.MPD.Value' to @String@ within a @Maybe@ context.
-
-This @Value@ is from 'Network.MPD' and is basically the same as a
-@String@ but used internally to store metadata values.
-
-==== __Example__:
-
-@
-processSong :: Metadata -> Maybe Song -> Maybe String
-processSong _ Nothing = Nothing
-processSong tag (Just song) = do
-  let tagVal = MPD.sgGetTag tag song
-  valueToStringMay =<< (headMay =<< tagVal)
-@
-
-'MPD.sgGetTag' returns a @Maybe [Value]@. [libmpd](Network.MPD) also provides
-'Network.MPD.toString' that can convert, along other types, a
-'Network.MPD.Value' to a @String@.
--}
-valueToStringMay :: MPD.Value -> Maybe String
-valueToStringMay x = Just (MPD.toString x)
-
-{- | Check if @Maybe v@ exists and is of type expected by
-'Data.Aeson.object' as defined in 'Data.Aeson.Value', if it is return
-both the @key@ and @value@ within the @Maybe@ context tied with
-'Data.Aeson..='. This gives support to \'optional\' fields using
-'Data.Maybe.catMaybes' that discard @Nothing@ values and is meant to
-prevent creating JSON key/value pairs with @null@ values, e.g.:
-
-@
-jsonTags = object . catMaybes $
-    [ "artist"  .=? artist
-    , "album"   .=? album
-    , "title"   .=? title
-    ]
-@
-
-Where if a value on the right is @Nothing@ that key/value pair will
-not be included in 'Data.Aeson.object' because of
-'Data.Maybe.catMaybes'.
--}
-(.=?) :: (KeyValue e a, ToJSON v) => Key -> Maybe v -> Maybe a
-key .=? Just value = Just (key .= value)
-_   .=? Nothing    = Nothing
-
--- | Helper function for creating an JSON 'Data.Aeson.object' where
--- 'Data.Maybe.catMaybes' won't include items from the '[Maybe Pair]'
--- list that return 'Nothing'.
-objectJson :: [Maybe Pair] -> Value
-objectJson = object . catMaybes
-
--- | Extracts the 'Int' value from an 'MPD.Id' within 'MPD.Status', if
--- present and the 'Either' value is a 'Right'.
+-- | Extracts the 'Int' value from an 'Network.MPD.Id' within
+-- 'Network.MPD.Status', if present and the 'Either' value is 'Right'.
 getStatusIdInt :: (MPD.Status -> Maybe MPD.Id) -> Either MPD.MPDError MPD.Status -> Maybe Int
 getStatusIdInt item status =
   case m of