+++ src/Options.hs
@@ -1,5 +1,6 @@
 module Options
   ( Opts(..)
+  , NextSongFlag(..)
   , execParser
   , prefs
   , showHelpOnEmpty
@@ -17,6 +18,7 @@ import Options.Applicative
       metavar,
       option,
       strOption,
+      flag',
       prefs,
       progDesc,
       short,
@@ -26,7 +28,9 @@ import Options.Applicative
       Parser,
       ParserInfo,
       infoOption,
-      hidden )
+      hidden,
+      many,
+      (<|>) )
 
 import Options.Applicative.Extra ( helperWith )
 
@@ -37,8 +41,13 @@ data Opts = Opts  -- ^ Custom data record for storing 'Options.Applicative.Parse
   { optPort    :: Integer  -- ^ MPD port to connect.
   , optHost    :: String   -- ^ MPD host address to connect.
   , optPass    :: String   -- ^ Plain text password to connect to MPD.
+  , optNext    :: NextSongFlag -- ^ Either include in the json or print it alone.
   , optVersion :: Type -> Type  -- ^ Print program version.
   }
+
+data NextSongFlag = IncludeNextSong
+                  | OnlyNextSong
+                  | NoNextSong
 
 optsParser :: Parser Opts
 optsParser
@@ -46,7 +55,11 @@ optsParser
   <$> portOptParser
   <*> hostOptParser
   <*> passOptParser
+  <*> nextSongOptParser
   <*> versionOptParse
+  where
+    nextSongOptParser = nextSongFlagCountOptParser
+                        <|> nextSongOnlyOptParser
 
 portOptParser :: Parser Integer
 portOptParser
@@ -74,6 +87,28 @@ passOptParser
   <> short 'P'
   <> value ""
   <> help "Password for connecting (will be sent as plain text)"
+
+nextSongFlagCountOptParser :: Parser NextSongFlag
+nextSongFlagCountOptParser =
+  fmap (intToNextSong . length) <$> many
+  $ flag' ()
+  $ short 'n'
+  <> long "next"
+  <> help ( concat
+            [ "If used once (e.g. -n), include next song information in the output.\n"
+            , "If used twice (e.g. -nn) it's an alias for --next-only." ])
+
+nextSongOnlyOptParser :: Parser NextSongFlag
+nextSongOnlyOptParser
+  = flag' OnlyNextSong
+    ( long "next-only"
+      <> help "Only print next song information." )
+
+intToNextSong :: Int -> NextSongFlag
+intToNextSong count
+  | count == 1 = IncludeNextSong
+  | count > 1 = OnlyNextSong
+  | otherwise = NoNextSong
 
 versionOptParse :: Parser (a -> a)
 versionOptParse =